+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - **Desktop:** Opening the macOS disk image now shows a drag-to-Applications installer window, with the application on the left and Applications on the right.
+- **Application:** Added a variety of early usability enhancements.
 
 ## [0.1.0] - 2026-09-03